Grid 1

+++ include/vecUtilities.h
@@ -6,4 +6,4 @@ std::vector<int> linspace(const int &ini, const int &fin, const int &n);
 std::vector<int> arange(const int &ini, const int &fin, const int &step);
 std::vector<int> zeros(const int &n);
 std::vector<int> ones(const int &n);
-void printV(const std::vector<int> &vec);
+void printV(const std::vector<double> &vec);
